% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_input_from_vectra.R
\name{make_input_from_vectra}
\alias{make_input_from_vectra}
\title{make_input_from_vectra}
\usage{
make_input_from_vectra(
  infile,
  tissue_category = c("Tumor", "Stroma"),
  compartment = c("Nucleus", "Entire.Cell", "Cytoplasm"),
  excluded_markers = c("AE1AE3", "DAPI")
)
}
\arguments{
\item{infile}{A tab delimited text file from Vectra machine}

\item{tissue_category}{What type of cells we are interested in. Options are "Tumor", "Stroma"}

\item{compartment}{Intensity values measured from which cellular compartment. Options are
'Nuleus', 'Entire.Cell' and 'Cytoplasm'}

\item{excluded_markers}{Marker(s) to exlude from the output. For example, we may not want 'AE1AE3' and/or 'DAPI'}
}
\value{
A data frame for use as input for the run_scTorus function
}
\description{
This function takes a consolidated data file from Vectra and output a data frame
as input for the function of run_scTorus for cell cycle trajectory analysis
and trajectory profiles showing the up- and downs of various cell cycle protine markers.
}
