% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scTorus.R
\name{run_scTorus}
\alias{run_scTorus}
\title{run_scTorus}
\usage{
run_scTorus(
  x,
  genes.to.highlight = NULL,
  method = c("lmds", "isoMDS", "CMD", "phate"),
  control.ROIs = NULL,
  treated.ROIs = NULL,
  smoothFactor = 5,
  reference.marker = NULL
)
}
\arguments{
\item{x}{A data frame with cell IDs as rows and markers as columns}

\item{genes.to.highlight}{Gene names to highlight using measured mean intensity values}

\item{method}{Dimension reduction method to use. Options are "lmds", "isoMDS", "CMD"}

\item{control.ROIs}{ROIs for cells in the control group}

\item{treated.ROIs}{ROIs for cells in the treated group}

\item{smoothFactor}{A integer values used to control the smoothness of the trajectory}
}
\value{
A list of ggplot2 plotting objects, including a scatterplot showing cells arranged as circles,
a heatmap showing the marker intensity along a calculated pseudo time, a line plot showing the dynamics
of intensity along the pseudo time
}
\description{
This function generate plotting objects for making cell cycle trajectory plots.
It will output a scatterplot showing single cells arranged in a circle and a heatmap
and trajectory profiles showing the up- and downs of various cell cycle protine markers.
}
